tests/wpt/tests/css/css-gaps/parsing/rule-edge-interior-inset-valid.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Gaps: parsing rule-inset shorthands with valid values</title>
  <link rel="author" title="Javier Contreras" href="mailto:javiercon@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#inset">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  <script>
    const properties = ['column-rule-edge-inset', 'row-rule-edge-inset', 'column-rule-interior-inset', 'row-rule-interior-inset', 'rule-edge-inset', 'rule-interior-inset'];
    for (const property of properties) {
      test_valid_value(property, '0', '0px');
      test_valid_value(property, '-20px', '-20px');
      test_valid_value(property, '5%', '5%');
      test_valid_value(property, '10vmin', '10vmin');
      test_valid_value(property, 'calc(8em + 4ex)', 'calc(8em + 4ex)');
      test_valid_value(property, '5% 10%', '5% 10%');
      test_valid_value(property, '-20px 10px', '-20px 10px');
    }
  </script>
</body>
</html>
